Removal of FixedChar260.

+++ src/engine/resources/AnimTileManager.h
@@ -20,6 +20,6 @@ const char* AnimTileManager_GetDatFileName();
 void AnimTileManager_LoadAllFromDat();
 void AnimTileManager_Dispose(const char* filenameWithoutExtension);
 void AnimTileManager_DisposeAll();
-int AnimTileManager_Length();
+int64_t AnimTileManager_Length();
 Resource* AnimTileManager_GetResourceByIndex(int index);
 AnimTile* AnimTileManager_GetResourceDataByIndex(int index);
